restore restock
